added background imaging and dashboard component created for a landing page, previous dashboard with the music search and player has been moved to its own section in music-player route
