+++ client/src/components/common/Layout.js
@@ -3,12 +3,14 @@ import React from "react";
 // import custom navBar and footer components to wrap application pages
 import NavBar from "./NavBar";
 import Footer from "./Footer";
+import ContentBackground from "./ContentBackground";
 
 export default function Layout({ children }) {
   return (
     <>
       <NavBar />
-      <div>{children}</div>
+      <ContentBackground style={{ margin: "60px 0 0 0" }} />
+      {children}
       <Footer />
     </>
   );
