+++ client/src/components/countdown/SongCountdown.js
@@ -64,76 +64,72 @@ export default function SongCountdown({
     console.log("Update in Players Array -> ", players);
   }, [players]);
 
-  return showModal !== "" ? (
-    showModal === "AddPlayerModal" ? (
-      <AddPlayerModal
-        showModal={showModal}
-        handleModalClose={handleModalClose}
-        addPlayer={addPlayer}
-        accessToken={accessToken}
-        editPlayerDetails={editPlayerDetails}
-      />
-    ) : (
-      <ImportPlaylistModal
-        showModal={showModal}
-        handleModalClose={handleModalClose}
-        accessToken={accessToken}
-      />
-    )
-  ) : (
-    <div
-      className="wrapper"
-      style={{
-        width: "100%",
-        height: "auto",
-        padding: "10px",
-        marginBottom: "110px",
-      }}
-    >
-      {!isGameStarted ? (
-        <Container
-          className="justify-content-center align-items-top py-50"
-          style={{
-            width: "70%",
-            minHeight: "100vh",
-            backgroundColor: "#f8f9fa",
-            marginBottom: "130px",
-            paddingTop: "10px",
-            paddingBottom: "10px",
-            borderRadius: "10px",
-          }}
-        >
-          <div className="position-relative top-20 start-20 mt-20">
-            <div className="d-flex gap-2">
-              <Button variant="primary" onClick={handleModalOpen}>
-                Add Player
-              </Button>
-              <Button variant="secondary" onClick={handleImportPlaylist}>
-                Import Playlist
-              </Button>
-              <Button variant="danger" onClick={startCountdown}>
-                Start Countdown
-              </Button>
-            </div>
-            <div className="grid-container mt-4">
-              {players.map((player) => (
-                <PlayerCard
-                  key={player.id}
-                  player={player}
-                  editPlayer={editPlayer}
-                  removePlayer={removePlayer}
-                />
-              ))}
-            </div>
-          </div>
-        </Container>
+  return (
+    <div className="content-container">
+      {showModal !== "" ? (
+        showModal === "AddPlayerModal" ? (
+          <AddPlayerModal
+            showModal={showModal}
+            handleModalClose={handleModalClose}
+            addPlayer={addPlayer}
+            accessToken={accessToken}
+            editPlayerDetails={editPlayerDetails}
+          />
+        ) : (
+          <ImportPlaylistModal
+            showModal={showModal}
+            handleModalClose={handleModalClose}
+            accessToken={accessToken}
+          />
+        )
       ) : (
-        <Hottest100Countdown
-          accessToken={accessToken}
-          players={players}
-          setPlayingTrack={setPlayingTrack}
-          playerRef={playerRef}
-        />
+        <>
+          {!isGameStarted ? (
+            <Container
+              className="justify-content-center align-items-top py-50"
+              style={{
+                width: "70%",
+                height: "auto",
+                backgroundColor: "#f8f9fa",
+                marginBottom: "130px",
+                paddingTop: "10px",
+                paddingBottom: "10px",
+                borderRadius: "10px",
+              }}
+            >
+              <div className="position-relative top-20 start-20 mt-20">
+                <div className="d-flex gap-2">
+                  <Button variant="primary" onClick={handleModalOpen}>
+                    Add Player
+                  </Button>
+                  <Button variant="secondary" onClick={handleImportPlaylist}>
+                    Import Playlist
+                  </Button>
+                  <Button variant="danger" onClick={startCountdown}>
+                    Start Countdown
+                  </Button>
+                </div>
+                <div className="grid-container mt-4">
+                  {players.map((player) => (
+                    <PlayerCard
+                      key={player.id}
+                      player={player}
+                      editPlayer={editPlayer}
+                      removePlayer={removePlayer}
+                    />
+                  ))}
+                </div>
+              </div>
+            </Container>
+          ) : (
+            <Hottest100Countdown
+              accessToken={accessToken}
+              players={players}
+              setPlayingTrack={setPlayingTrack}
+              playerRef={playerRef}
+            />
+          )}
+        </>
       )}
     </div>
   );
