+++ client/src/components/dashboard/Dashboard.js
@@ -1,83 +1,11 @@
 import { useState, useEffect } from "react";
 import { Container, Form } from "react-bootstrap";
-import SpotifyWebApi from "spotify-web-api-node";
+import "./css/dashboard.css";
 
-// Import custom functional components
-import TrackSearchResult from "../common/TrackSearchResult";
-
-const spotifyApi = new SpotifyWebApi({
-  clientId: "af336f24a30e439b88ed899c0813426a",
-});
-
-export default function Dashboard({ accessToken, setPlayingTrack }) {
-  const [search, setSearch] = useState("");
-  const [searchResults, setSearchResults] = useState([]);
-
-  function chooseTrack(track) {
-    setPlayingTrack(track);
-    setSearch("");
-  }
-
-  useEffect(() => {
-    if (!accessToken) {
-      return;
-    }
-    spotifyApi.setAccessToken(accessToken);
-  }, [accessToken]);
-
-  useEffect(() => {
-    if (!search) {
-      return setSearchResults([]);
-    } // if there isn't a value for search then set results to empty array
-    if (!accessToken) {
-      return;
-    }
-
-    // search spotify based on search value
-    let cancel = false;
-    spotifyApi.searchTracks(search).then((res) => {
-      if (cancel) return;
-      setSearchResults(
-        res.body.tracks.items.map((track) => {
-          const smallestImage = track.album.images.reduce((smallest, image) => {
-            if (image.height < smallest.height) return image;
-            return smallest;
-          });
-          return {
-            artist: track.artists[0].name,
-            title: track.name,
-            uri: track.uri,
-            albumUrl: smallestImage.url,
-          };
-        })
-      );
-      console.log(res);
-    });
-    return () => (cancel = true);
-  }, [search, accessToken]);
-
+export default function Dashboard() {
   return (
-    <div>
-      <Container
-        className="d-flex flex-column py-2"
-        style={{ height: "100vh" }}
-      >
-        <Form.Control
-          type="search"
-          placeholder="search songs/artist"
-          value={search}
-          onChange={(e) => setSearch(e.target.value)}
-        />
-        <div className="flex-grow-1 my-2" style={{ overflowY: "auto" }}>
-          {searchResults.map((track) => (
-            <TrackSearchResult
-              track={track}
-              key={track.uri}
-              chooseTrack={chooseTrack}
-            />
-          ))}
-        </div>
-      </Container>
-    </div>
+    <>
+      <div className="page-title">SOUND-WAVE</div>
+    </>
   );
 }
