+++ client/src/components/musicplayer/css/MusicPlayer.css
@@ -1,0 +1,6 @@
+.music-player-container {
+  background-color: white;
+  width: 50%;
+  padding: 10px;
+  border-radius: 10px;
+}
