+++ client/src/global.css
@@ -1,0 +1,24 @@
+.content-container {
+  width: 100%;
+  height: auto;
+  position: absolute;
+  top: 70px;
+  display: flex;
+  justify-content: center;
+}
+
+.custom-modal .modal-dialog {
+  position: absolute;
+  width: 100%;
+  left: 0;
+  right: 0;
+  margin-left: auto;
+  margin-right: auto;
+  max-width: 1000px;
+  z-index: 1050;
+}
+
+.custom-modal .modal-content {
+  border-radius: 8px;
+  padding: 20px;
+}
